Created app.js
Created the initial javascript file.

Included are 2 sets of functions. The first three functions affect the visibility of the login entry popup.

The following function is the second custom feature of the project:
On scroll the navigation bar animates, reducing text size and becomes sticky, staying at the top of the viewport. The navBar has a return animation whenthe viewport scrolls back to the top of the page.